REFACT: GameSceneManager
